fix(sm): Domain update collisions still happening
Saw a recurrence of previously-observed intermittent issue where
multiple constructs can attempt to call sagemaker:UpdateDomain
concurrently in stack creation. Factor the previous fix for this
out into a separate Lambda layer to also bring it to another
previously-uncovered resource, and try broadening the error catch to
see if that's what was failing.