build(cdk): Install cdk-nag (disabled for now)
Install the cdk-nag tool and add a note to the readme explaniing how
it can be used to enumerate outstanding security config changes
suggested before taking the stack to production. Disabled for now
because we'd have to enumerate & ignore a lot of IAM permission
findings to run the build successfully with it turned on.